Update IndexBehaviour.entry/1 return typespec
This now matches the documentation for the `entry/1` callback. Using a keyword literal for typing requires that the keys are the atoms that will be returned by the function. i.e. – before this change, it was expecting a keyword list in this shape:

    [key: "a string"]

or

    [key: {:fa, "hand-wave"}]

Now either key is optional to dialyzer and their values are typed.